0-print_list.c

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -10,12 +10,12 @@
 
 size_t print_list(const list_t *h)
 {
-	size_t count = 0;
+	int count = 0;
 
 	while (h != NULL)
 	{
 		if (h->str != NULL)
-			printf("[%u] %s\n", h->len, h->str);
+			printf("[%d] %s\n", h->len, h->str);
 		else
 			printf("[0] (nil)\n");
